feat : 생성일자 컬럼 추가 및 관련 코드 수정
