+++ src/main/java/com/nhnacademy/bookstoreuserapi/common/controller/advice/ErrorMessage.java
@@ -3,8 +3,8 @@ package com.nhnacademy.bookstoreuserapi.common.controller.advice;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 
-import java.time.ZoneOffset;
-import java.time.ZonedDateTime;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 
 @Getter
 @Schema(
@@ -13,7 +13,7 @@ import java.time.ZonedDateTime;
         name = "ErrorMessage")
 public class ErrorMessage {
     @Schema(description = "에러 발생 시간")
-    private final String timeStamp;
+    private final String timestamp;
     @Schema(description = "HTTP 상태 코드")
     private final int status;
     @Schema(description = "에러 메시지")
@@ -24,7 +24,7 @@ public class ErrorMessage {
     private final String message;
 
     public ErrorMessage(int status, String error, String path, String message) {
-        this.timeStamp = ZonedDateTime.now(ZoneOffset.UTC).toString();
+        this.timestamp = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
         this.status = status;
         this.error = error;
         this.path = path;
