+++ src/test/java/com/nhnacademy/bookstoreuserapi/point/repository/PointRepositoryTest.java
@@ -1,0 +1,101 @@
+package com.nhnacademy.bookstoreuserapi.point.repository;
+
+import com.nhnacademy.bookstoreuserapi.common.config.QuerydslConfig;
+import com.nhnacademy.bookstoreuserapi.point.domain.Point;
+import com.nhnacademy.bookstoreuserapi.pointtype.domain.PointType;
+import com.nhnacademy.bookstoreuserapi.pointtype.repository.PointTypeRepository;
+import com.nhnacademy.bookstoreuserapi.user.domain.User;
+import com.nhnacademy.bookstoreuserapi.user.domain.User.Status;
+import com.nhnacademy.bookstoreuserapi.usergrade.domain.UserGrade;
+import com.nhnacademy.bookstoreuserapi.point.domain.ResponsePoint;
+import com.nhnacademy.bookstoreuserapi.user.repository.UserRepository;
+import com.nhnacademy.bookstoreuserapi.usergrade.repository.UserGradeRepository;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.context.annotation.Import;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import static com.nhnacademy.bookstoreuserapi.usergrade.domain.UserGrade.Grade.BASIC;
+import static org.assertj.core.api.Assertions.assertThat;
+
+@DataJpaTest
+@ActiveProfiles("test")
+@Import(QuerydslConfig.class)
+class PointRepositoryTest {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
+    private PointRepository pointRepository;
+
+    @Autowired
+    private UserGradeRepository userGradeRepository;
+
+    @Autowired
+    private PointTypeRepository pointTypeRepository;
+
+    private final String userId = "testUser";
+
+    @BeforeEach
+    void setUp() {
+        User user = new User(
+                userId,
+                "password",
+                "홍길동",
+                "01012345678",
+                "hong@test.com",
+                java.time.LocalDate.of(1995, 5, 1)
+        );
+        user.setUserStatus(Status.ACTIVE);
+        user.setUserPoint(100);
+        user.setCreatedAt(LocalDateTime.now());
+
+        UserGrade userGrade = new UserGrade(BASIC,0L);
+
+        userGradeRepository.save(userGrade);
+
+        user.setUserGrade(userGrade);
+
+        userRepository.save(user);
+
+        PointType pointType1 = new PointType(null, "회원가입", 5000, 1, userGrade);
+        PointType pointType2 = new PointType(null,"리뷰작성", 500, 1, userGrade);
+
+        pointTypeRepository.save(pointType1);
+        pointTypeRepository.save(pointType2);
+
+        Point point1 = new Point(null, user, pointType1, 1L, LocalDateTime.now(), 50);
+        Point point2 = new Point(null, user, pointType2, 2L, LocalDateTime.now(), -100);
+
+        pointRepository.save(point1);
+        pointRepository.save(point2);
+    }
+
+    @Test
+    @DisplayName("사용자 ID로 포인트 조회")
+    void testFindPointByUserId() {
+        Pageable pageable = PageRequest.of(0, 10);
+
+        List<ResponsePoint> points = pointRepository.findPointByUserId(userId, pageable).getContent();
+
+        assertThat(points).hasSize(2);
+        assertThat(points)
+                .extracting("earnedAndUsedPoint")
+                .containsExactlyInAnyOrder(50, -100);
+    }
+
+    @AfterEach
+    void tearDown() {
+        pointRepository.deleteAll();
+    }
+}
